[FEATURE] Upgrade for PHP, Sentry SDK, TYPO3
This comes with a major refactoring and cleanup
Everything can be fully configured using TYPO3 configuration
or Dependency Injection Service configuration.

Documentation still applies, but especially design decisions,
what exactly is mapped from TYPO3 to a Sentry event
needs to be done.

Also practical examples should be added to the documentation,
and of course developer and integrator APIs that are now provided
need some explanation.