[BUGFIX] Also handle deprecation messages
Deprecation messages are logged to a specific log target.
In order to make it possible to also log these cleanly into Sentry,
the error handler needs to check for them explicitly.

To be fully aligned with how TYPO3 Console logs errors,
and to have more descriptive exception names in Sentry,
the package helhum/php-error-reporting is now added
as requirement and used in error handler.